Adds interface
