+++ app/src/main/java/com/thedistantblue/triaryapp/utils/AutoDisposable.java
@@ -2,26 +2,10 @@ package com.thedistantblue.triaryapp.utils;
 
 import androidx.lifecycle.Lifecycle;
 import androidx.lifecycle.LifecycleObserver;
-import androidx.lifecycle.OnLifecycleEvent;
 
-import io.reactivex.rxjava3.disposables.CompositeDisposable;
 import io.reactivex.rxjava3.disposables.Disposable;
 
-public class AutoDisposable implements LifecycleObserver {
-
-    private CompositeDisposable compositeDisposable;
-
-    public void bindTo(Lifecycle lifecycle) {
-        lifecycle.addObserver(this);
-        compositeDisposable = new CompositeDisposable();
-    }
-
-    public void add(Disposable disposable) {
-        compositeDisposable.add(disposable);
-    }
-
-    @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
-    public void onDestroy() {
-        compositeDisposable.clear();
-    }
+public interface AutoDisposable extends LifecycleObserver {
+    void bindTo(Lifecycle lifecycle);
+    void add(Disposable disposable);
 }
