+++ app/src/main/java/com/thedistantblue/triaryapp/utils/AutoDisposableImpl.java
@@ -1,0 +1,28 @@
+package com.thedistantblue.triaryapp.utils;
+
+import androidx.lifecycle.Lifecycle;
+import androidx.lifecycle.OnLifecycleEvent;
+
+import io.reactivex.rxjava3.disposables.CompositeDisposable;
+import io.reactivex.rxjava3.disposables.Disposable;
+
+public class AutoDisposableImpl implements AutoDisposable {
+
+    private CompositeDisposable compositeDisposable;
+
+    @Override
+    public void bindTo(Lifecycle lifecycle) {
+        lifecycle.addObserver(this);
+        compositeDisposable = new CompositeDisposable();
+    }
+
+    @Override
+    public void add(Disposable disposable) {
+        compositeDisposable.add(disposable);
+    }
+
+    @OnLifecycleEvent(Lifecycle.Event.ON_DESTROY)
+    public void onDestroy() {
+        compositeDisposable.clear();
+    }
+}
